+++ sdk/net/basic/server_socket.hpp
@@ -91,11 +91,16 @@ class server_socket {
      * @param   char*           ptr_data        [OUT]
      * @param   size_t          size_data       [IN]
      * @param   size_t*         cbread          [OUT]
+     * @param   struct sockaddr* addr           [outopt]
+     * @param   socklen_t* addrlen              [inopt]
      * @return  error code (see error.hpp)
      * @remarks
      *          ERROR_CONNECTION_CLOSED
      */
-    virtual return_t read(socket_t sock, tls_context_t* tls_handle, int mode, char* ptr_data, size_t size_data, size_t* cbread) { return errorcode_t::success; }
+    virtual return_t read(socket_t sock, tls_context_t* tls_handle, int mode, char* ptr_data, size_t size_data, size_t* cbread, struct sockaddr* addr = nullptr,
+                          socklen_t* addrlen = nullptr) {
+        return errorcode_t::success;
+    }
     /**
      * @brief   send
      * @param   socket_t        sock            [IN]