+++ sdk/net/basic/tcp_server_socket.cpp
@@ -64,7 +64,8 @@ return_t tcp_server_socket::accept(socket_t sock, socket_t* clisock, struct sock
     return ret;
 }
 
-return_t tcp_server_socket::read(socket_t sock, tls_context_t* tls_handle, int mode, char* ptr_data, size_t size_data, size_t* cbread) {
+return_t tcp_server_socket::read(socket_t sock, tls_context_t* tls_handle, int mode, char* ptr_data, size_t size_data, size_t* cbread, struct sockaddr* addr,
+                                 socklen_t* addrlen) {
     return_t ret = errorcode_t::success;
 
     __try2 {